add int8 test framework

+++ src/HowToOptimizeGemm/REF_MMult.c
@@ -5,17 +5,18 @@
 #define C(i,j) c[ (i)*ldc + (j) ]
 
 /* Routine for computing C = A * B + C */
+#include <stdlib.h>
 
-void REF_MMult( int m, int n, int k, float *a, int lda, 
-                                    float *b, int ldb,
-                                    float *c, int ldc )
+void REF_MMult( int m, int n, int k, int8_t *a, int lda, 
+                                     int8_t *b, int ldb,
+                                     int32_t *c, int ldc )
 {
   int i, j, p;
 
   for ( i=0; i<m; i++ ){
     for ( j=0; j<n; j++ ){
       for ( p=0; p<k; p++ ){
-	C( i,j ) = C( i,j ) +  A( i,p ) * B( p,j );
+        C( i,j ) = C( i,j ) +  A( i,p ) * B( p,j );
       }
     }
   }
